synaptics: fix touchscreen getting stuck with h2w/l2w/s2w
The prevention of touch events being synced when screen is
off with wake irq caused unwanted effects of occasional
touchscreen events not syncing properly when screen was
switched on again (when some touch events were left
without a second event of finger left).

Now instead of not syncing the input events when
wake option is keeping irq wake, we sync the finger
events but with a false coordinate outside the screen.
This way while screen is off (either power button sleep
or in call proximity sensor triggered), it won't send
events to framework in the screen area, fixing both
gettings stuck and unwanted touch events in calls/sleep.

Change-Id: I699f4b0561ebd2b409b8afe61acf820ac9a9989d